backend/requirements.txt
protobuf<=3.20.1
motor==2.5.1
uvicorn[standard]==0.15.0
ortools==9.0.9048
tabulate==0.8.9
pydantic==1.9.0
pandas==1.3.3
dateparser==1.0.0
numpy==1.21.2
beanie==1.7.2
fastapi==0.68.1
python-dotenv==0.19.0
icalendar==4.0.7
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
fastapi_permissions==0.2.7
fire==0.4.0
msal==1.17.0
